feat: add options to compute_report
Now it is possible to group by exercise or not, by year, month, week or nothing.
The report will automatically be sorted in a way that makes sense.

TODO: test this, check that the values are correct.
TODO: use this method only, remove similar methods in Exercise and Workout models.