test: test compute yearly per exercise report
TODO: before testing the other compute_report settings, refactor the existing tests. There should be a parametrized fixture that returns file, dataframe, start and end dates. And a function (or other fixture) that computes the report from the dataframe, based on the same parameters passed to the compute_report method.