feat: wip, default muscle group logic
TODO: this shouldn't be done with signals: they complicate things too much.
Also, the m2m_changed signal is not triggered by the admin interface.

See this discussion: https://forum.djangoproject.com/t/how-to-set-default-value-to-manytomanyfield/27971

Basically,  in this case it is recommended to define a custom view. But I really like the Admin, so try to use the Admin as "view", and add logic to its save_model, save_related methods.

Doing this in the save method of Exercise is (probably) impossible, since the method is not aware of the changes in the related model. It is only aware of the current number of relations, but not how the relations are presently being changed.
Admin save_related, on the other hand, should be aware of the relations being changed.

So, in short, TODO:
try to implement this after the super().save_related call in a custom save_related admin method.