feat: make muscle groups of exercise logic work
Now the default muscle group is added when there are no muscle groups available.
If all are removed, the default is added back.
If there are more than one muscle group, the default is removed.

I decided to do this in the Admin.
In this forum question:
https://forum.djangoproject.com/t/how-to-set-default-value-to-manytomanyfield/27971
it was suggested that I didn't use the admin and created a custom view.
Also, the pros and cons of signals and custom model save methods were discussed.

In the end I decided that:
- signals would increase complexity a lot for this simple logic
- it was not possible to add this behaviour in a single save method. I should have updated more than one save method, like of exercise and musclegroups. Also, with custom save methods, the logic would be skipped if alternative save methods were used (like bulk save).

In the end I decided that the custom view was a good approach, at least for simplicity of implementation. But since I didn't want to implement a custom view, because I like the Admin interface, I decided that the closest thing to implementing a custom view would be to override the save_related method.

NOTE: make sure that all Exercises are created via the Admin! If additional Exercise creation views were added, the same logic must be applied there, too!!!